eliminate the pervasive capture of `self` in notification blocks
The previous approach of converting closures to dispatch_block_t early came from an in initial attempt that used dispatch queues as storage.
Since blocks are now only dispatched after a `Deferred` has become determined, it has become unnecessary to store a back reference in every block, since we know the value of `result` at the time blocks are dispatched.

The new approach stores notification handlers and qos values; they are dispatched with a copy of result and no capture of `self`.
This way, a `Deferred` could be deallocated before its last notification starts processing; that couldn't be the case previously.